[BREAKING] Move the Lambda Invoke CodePipeline Action.
This commit moves the Lambda Invoke Action from the codepipeline module to its dedicated,
new lambda-codepipeline module.
It's not possible to move it to the lambda module itself,
as that would cause a cyclic dependency between the codepipeline,
sns and lambda modules.

This continues the refactoring of moving out concrete Action typesfrom the codepipeline module.
