+++ packages/@aws-cdk/aws-codepipeline/lib/actions.ts
@@ -1,5 +1,4 @@
 import events = require('@aws-cdk/aws-events');
-import lambda = require('@aws-cdk/aws-lambda');
 import s3 = require('@aws-cdk/aws-s3');
 import cdk = require('@aws-cdk/cdk');
 import { Artifact } from './artifact';
@@ -30,7 +29,7 @@ export interface ActionArtifactBounds {
     maxOutputs: number;
 }
 
-function defaultBounds(): ActionArtifactBounds {
+export function defaultBounds(): ActionArtifactBounds {
     return {
         minInputs: 0,
         maxInputs: 5,
@@ -439,84 +438,6 @@ export class ApprovalAction extends Action {
             provider: 'Manual',
             artifactBounds: { minInputs: 0, maxInputs: 0, minOutputs: 0, maxOutputs: 0 }
         });
-    }
-}
-
-export interface InvokeLambdaProps {
-    /**
-     * The lambda function to invoke.
-     */
-    lambda: lambda.LambdaRef;
-
-    /**
-     * String to be used in the event data parameter passed to the Lambda
-     * function
-     *
-     * See an example JSON event in the CodePipeline documentation.
-     *
-     * https://docs.aws.amazon.com/codepipeline/latest/userguide/actions-invoke-lambda-function.html#actions-invoke-lambda-function-json-event-example
-     */
-    userParameters?: any;
-
-    /**
-     * Adds the "codepipeline:PutJobSuccessResult" and
-     * "codepipeline:PutJobFailureResult" for '*' resource to the Lambda
-     * execution role policy.
-     *
-     * NOTE: the reason we can't add the specific pipeline ARN as a resource is
-     * to avoid a cyclic dependency between the pipeline and the Lambda function
-     * (the pipeline references) the Lambda and the Lambda needs permissions on
-     * the pipeline.
-     *
-     * @see
-     * https://docs.aws.amazon.com/codepipeline/latest/userguide/actions-invoke-lambda-function.html#actions-invoke-lambda-function-create-function
-     *
-     * @default true
-     */
-    addPutJobResultPolicy?: boolean;
-}
-/**
- * @link https://docs.aws.amazon.com/codepipeline/latest/userguide/actions-invoke-lambda-function.html
- */
-export class InvokeLambdaAction extends Action {
-    constructor(parent: Stage, name: string, props: InvokeLambdaProps) {
-        super(parent, name, {
-            category: ActionCategory.Invoke,
-            provider: 'Lambda',
-            artifactBounds: defaultBounds(),
-            configuration: {
-                FunctionName: props.lambda.functionName,
-                UserParameters: props.userParameters
-            }
-        });
-
-        // allow pipeline to list functions
-        parent.pipeline.addToRolePolicy(new cdk.PolicyStatement()
-            .addAction('lambda:ListFunctions')
-            .addResource('*'));
-
-        // allow pipeline to invoke this lambda functionn
-        parent.pipeline.addToRolePolicy(new cdk.PolicyStatement()
-            .addAction('lambda:InvokeFunction')
-            .addResource(props.lambda.functionArn));
-
-        // allow lambda to put job results for this pipeline.
-        const addToPolicy = props.addPutJobResultPolicy !== undefined ? props.addPutJobResultPolicy : true;
-        if (addToPolicy) {
-            props.lambda.addToRolePolicy(new cdk.PolicyStatement()
-                .addResource('*') // to avoid cycles (see docs)
-                .addAction('codepipeline:PutJobSuccessResult')
-                .addAction('codepipeline:PutJobFailureResult'));
-        }
-    }
-
-    /**
-     * Add an input artifact
-     * @param artifact
-     */
-    protected addInputArtifact(artifact: Artifact): Action {
-        super.addInputArtifact(artifact);
-        return this;
     }
 }
