+++ packages/@aws-cdk/aws-codepipeline/test/test.action.ts
@@ -1,5 +1,3 @@
-import { expect, haveResource } from '@aws-cdk/assert';
-import lambda = require('@aws-cdk/aws-lambda');
 import cdk = require('@aws-cdk/cdk');
 import { Test } from 'nodeunit';
 import codepipeline = require('../lib');
@@ -86,87 +84,6 @@ export = {
             outputArtifacts: [{ name: 'TestOutput' }],
             runOrder: 1
         });
-        test.done();
-    },
-
-    'InvokeLambdaAction can be used to invoke lambda functions from a pipeline'(test: Test) {
-        const stack = new cdk.Stack();
-
-        const fction = new lambda.Lambda(stack, 'Function', {
-            code: new lambda.LambdaInlineCode('bla'),
-            handler: 'index.handler',
-            runtime: lambda.LambdaRuntime.NodeJS43,
-        });
-
-        const pipeline = new codepipeline.Pipeline(stack, 'Pipeline');
-
-        new codepipeline.InvokeLambdaAction(new codepipeline.Stage(pipeline, 'Stage'), 'InvokeAction', {
-            lambda: fction,
-            userParameters: 'foo-bar/42'
-        });
-
-        expect(stack).to(haveResource('AWS::CodePipeline::Pipeline', {
-            "ArtifactStore": {
-              "Location": {
-                "Ref": "PipelineArtifactsBucket22248F97"
-              },
-              "Type": "S3"
-            },
-            "RoleArn": {
-              "Fn::GetAtt": [
-                "PipelineRoleD68726F7",
-                "Arn"
-              ]
-            },
-            "Stages": [
-              {
-                "Actions": [
-                  {
-                    "ActionTypeId": {
-                      "Category": "Invoke",
-                      "Owner": "AWS",
-                      "Provider": "Lambda",
-                      "Version": "1"
-                    },
-                    "Configuration": {
-                      "FunctionName": {
-                        "Ref": "Function76856677"
-                      },
-                      "UserParameters": "foo-bar/42"
-                    },
-                    "InputArtifacts": [],
-                    "Name": "InvokeAction",
-                    "OutputArtifacts": [],
-                    "RunOrder": 1
-                  }
-                ],
-                "Name": "Stage"
-              }
-            ]
-        }));
-
-        expect(stack).to(haveResource('AWS::IAM::Policy', {
-            "PolicyDocument": {
-              "Statement": [
-                {
-                  "Action": [
-                    "codepipeline:PutJobSuccessResult",
-                    "codepipeline:PutJobFailureResult"
-                  ],
-                  "Effect": "Allow",
-                  "Resource": "*"
-                }
-              ],
-              "Version": "2012-10-17"
-            },
-            "PolicyName": "FunctionServiceRoleDefaultPolicy2F49994A",
-            "Roles": [
-              {
-                "Ref": "FunctionServiceRole675BB04A"
-              }
-            ]
-        }));
-
         test.done();
     }
 };
