+++ packages/@aws-cdk/aws-lambda-codepipeline/.gitignore
@@ -1,0 +1,15 @@
+*.js
+*.js.map
+*.d.ts
+node_modules
+*.generated.ts
+dist
+tsconfig.json
+tslint.json
+
+.jsii
+
+.LAST_BUILD
+.nyc_output
+coverage
+.nycrc
