+++ packages/@aws-cdk/aws-lambda-codepipeline/.npmignore
@@ -1,0 +1,6 @@
+# Don't include original .ts files when doing `npm pack`
+*.ts
+!*.d.ts
+coverage
+.nyc_output
+*.tgz
