+++ packages/@aws-cdk/aws-lambda-codepipeline/lib/index.ts
@@ -1,0 +1,1 @@
+export * from './pipeline-action';
