+++ packages/@aws-cdk/aws-lambda-codepipeline/package.json
@@ -1,21 +1,21 @@
 {
-  "name": "@aws-cdk/aws-codepipeline",
+  "name": "@aws-cdk/aws-lambda-codepipeline",
   "version": "0.7.3-beta",
-  "description": "Better interface to AWS Code Pipeline",
+  "description": "AWS CodePipline Actions for AWS Lambda",
   "main": "lib/index.js",
   "types": "lib/index.d.ts",
   "jsii": {
     "outdir": "dist",
     "targets": {
       "java": {
-        "package": "com.amazonaws.cdk.aws.codepipeline",
+        "package": "com.amazonaws.cdk.aws.lambdacodepipeline",
         "maven": {
           "groupId": "com.amazonaws.cdk",
-          "artifactId": "aws-codepipeline"
+          "artifactId": "aws-lambdacodepipeline"
         }
       },
       "dotnet": {
-        "namespace": "Amazon.CDK.AWS.CodePipeline"
+        "namespace": "Amazon.CDK.AWS.LambdaCodePipeline"
       }
     }
   },
@@ -31,18 +31,16 @@
     "integ": "cdk-integ",
     "pkglint": "pkglint -f"
   },
-  "cdk-build": {
-    "cloudformation": "AWS::CodePipeline"
-  },
   "nyc": {
-    "lines": 66
+    "lines": 60,
+    "branches": 30
   },
   "keywords": [
     "aws",
     "cdk",
-    "cloudlib",
-    "aws-cloudlib",
-    "aws-clib"
+    "codepipeline",
+    "constructs",
+    "lambda"
   ],
   "author": {
     "name": "Amazon Web Services",
@@ -50,21 +48,15 @@
   },
   "license": "LicenseRef-LICENSE",
   "devDependencies": {
-    "@aws-cdk/aws-sns": "^0.7.3-beta",
     "@aws-cdk/assert": "^0.7.3-beta",
     "cdk-build-tools": "^0.7.3-beta",
     "cdk-integ-tools": "^0.7.3-beta",
-    "cfn2ts": "^0.7.3-beta",
     "pkglint": "^0.7.3-beta"
   },
   "dependencies": {
-    "@aws-cdk/aws-events": "^0.7.3-beta",
-    "@aws-cdk/aws-iam": "^0.7.3-beta",
+    "@aws-cdk/aws-codepipeline": "^0.7.3-beta",
+    "@aws-cdk/aws-lambda": "^0.7.3-beta",
     "@aws-cdk/aws-s3": "^0.7.3-beta",
-    "@aws-cdk/cdk": "^0.7.3-beta",
-    "@aws-cdk/util": "^0.7.3-beta"
-  },
-  "bundledDependencies": [
-    "@aws-cdk/util"
-  ]
+    "@aws-cdk/cdk": "^0.7.3-beta"
+  }
 }
